requirements.txt
pandas
numpy
streamlit
Pillow
opencv-python-headless
